specific_consumer/requirements.txt
nats-py==2.2.0
protobuf==4.21.9
